quotesbackend/requirements.txt
django==3.2
django-cors-headers 
django-taggit 
djangorestframework ==3.13.1
beautifulsoup4 
requests